Make network.Conn have the same id locally and on the remote server
When using network.DialTCP, the returned network.Conn has the same ID that is used on
the server that was dialed. This is accomplished by the server sending
the ID as the first message, and network.DialTCP trying to receive a
message and parsing the ID, before returning the connection.